+++ src/components/Button/AddClothesButton/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 import { Link } from 'react-router-dom';
 import styled from 'styled-components';
-import { ReactComponent as Plus } from '../../assets/plus.svg';
+import { ReactComponent as Plus } from '../../../assets/plus.svg';
 
 const AddClothesButton: React.FC = () => {
   return (
