+++ src/pages/MainPage/index.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
-import WeatherCard from '../MainPage/WeatherCard/WeatherCard';
-import MainList from './MainList/MainList';
+import WeatherCard from '../../components/WeatherCard/index';
+import MainList from '../../components/MainList/index';
 import { ReactComponent as Cloud } from '../../assets/weathers/cloud.svg';
 import { ReactComponent as Sun } from '../../assets/weathers/sun.svg';
 import { PageLayout, MainListWrapper, ProfileBox, TextBox, HeaderBox, NoticeWrapper, ProfileIconWrapper } from './style';
